refactor(simulation): consolidate RoboCasa examples
