+++ examples/advanced/robocasa/__init__.py
@@ -1,0 +1,1 @@
+"""Self-contained Retriever examples for the RoboSuite-to-RoboCasa stack."""
