+++ examples/advanced/robocasa/robosuite_inspection.py
