--- examples/advanced/robocasa_replay/__init__.py
@@ -1,1 +1,0 @@
-"""Retriever-connected RoboCasa demonstration replay example."""
