--- examples/advanced/robosuite/README.md
@@ -1,14 +1,0 @@
-# robosuite examples
-
-Work in progress!
-
-```bash
-pixi install -e robosuite
-pixi shell -e robosuite
-```
-
-Once in the shell, do the setup steps
-
-```bash
-python examples/advanced/robosuite/robosuite_example.py 
-```
