--- examples/advanced/robosuite_lift/__init__.py
@@ -1,1 +1,0 @@
-"""Optional robosuite integration demo."""
